undo move and delte facade

+++ src/pieces/models/bishop/bishop.cpp
@@ -2,7 +2,7 @@
 #include "../../factory/piece_factory.hpp"
 
 std::shared_ptr<Piece> Bishop::clone() {
-    std::shared_ptr<Piece> piece = color == PieceColor::black ? Piece::Create().Bishop.black().Position(position.x, position.y) : Piece::Create().Bishop.white().Position(position.x, position.y);
+    std::shared_ptr<Piece> piece = color == PieceColor::black ? Piece::Create().Bishop.black().Position(static_cast<int>(position.x), static_cast<int>(position.y)) : Piece::Create().Bishop.white().Position(static_cast<int>(position.x), static_cast<int>(position.y));
     piece->legalMoves = legalMoves;
     return piece;
 }
